+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Provider } from "react-redux"
 import "./App.css"
 import { store } from "./redux/store"
 import { setupAxiosInstance } from "./utils/axios-instance"
-import { PersistLogin } from "./persist-login"
+import { PersistLogin } from "./features/persist-login"
 
 function App() {
   setupAxiosInstance(store)
